api/pyproject.toml
[project]
name = "url-shortener-api"
version = "0.1.0"
description = "URL Shortener API"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "psycopg2-binary==2.9.9",
    "redis==5.0.1",
    "pydantic==2.5.0",
    "prometheus-client==0.19.0",
]
